Re #2407 Move useSHA out of Stack.Types.Config
